move package and rename to FpsMeasureUtil
