+++ app/src/main/java/io/github/droidkaigi/confsched2017/view/activity/SplashActivity.java
@@ -11,9 +11,9 @@ import java.util.concurrent.TimeUnit;
 import javax.inject.Inject;
 
 import io.github.droidkaigi.confsched2017.R;
-import io.github.droidkaigi.confsched2017.fps.FpsMeasure;
 import io.github.droidkaigi.confsched2017.repository.sessions.MySessionsRepository;
 import io.github.droidkaigi.confsched2017.repository.sessions.SessionsRepository;
+import io.github.droidkaigi.confsched2017.util.FpsMeasureUtil;
 import io.reactivex.Observable;
 import io.reactivex.Single;
 import io.reactivex.android.schedulers.AndroidSchedulers;
@@ -55,7 +55,7 @@ public class SplashActivity extends BaseActivity {
         loadSessionsForCache();
 
         // Starting new Activity normally will not destroy this Activity, so set this up in start/stop cycle
-        FpsMeasure.play(getApplication());
+        FpsMeasureUtil.play(getApplication());
     }
 
     @Override
@@ -64,7 +64,7 @@ public class SplashActivity extends BaseActivity {
         compositeDisposable.dispose();
 
         // Stop tracking the frame rate.
-        FpsMeasure.finish();
+        FpsMeasureUtil.finish();
     }
 
     private void loadSessionsForCache() {
