+++ app/src/release/java/io/github/droidkaigi/confsched2017/util/FpsMeasureUtil.java
@@ -1,10 +1,10 @@
-package io.github.droidkaigi.confsched2017.fps;
+package io.github.droidkaigi.confsched2017.util;
 
 import android.app.Application;
 
-public class FpsMeasure {
+public class FpsMeasureUtil {
 
-    private FpsMeasure() {
+    private FpsMeasureUtil() {
     }
 
     public static void play(Application application) {
